pyproject.toml
[project]
name = "mpf-employer-rag"
version = "0.1.0"
description = "Agentic RAG with DeepAgents + Zvec + MCP"
requires-python = ">=3.11"
dependencies = [
    "zvec>=0.3.0",
    "deepagents>=0.1.0",
    "langchain-openai>=0.2.0",
    "langchain-core>=0.3.0",
    "langgraph>=0.2.0",
    "sentence-transformers>=3.0.0",
    "fastmcp>=0.1.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "tiktoken>=0.7.0",
    "markdown>=3.0.0",
    "langchain-text-splitters>=1.1.1",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["mcp*", "rag*", "agents*", "config*"]
